RW-190 fixed test cases to use constructor of class CohortBuilderController.
